cambios del 15 05 2019

+++ Java-SpringBoot-WebService/Java-SpringBoot-Rest/src/main/java/com/example/domain/children/ChildrenReposytoryInterface.java
@@ -9,9 +9,9 @@ public interface ChildrenReposytoryInterface {
 
     public void actualizar(Children data);
 
-    public void eliminar(int key);
+    public void eliminar(long key);
 
-    public Children consultar(int key);
+    public Children consultar(long key);
 
     public List<Children> consultarTodoList();
 }
